Finished Libraries, built testbed
Added MD5 and Lookup3 to the libraries. All functions have been more or less standardized to a uint32_t output (array of uint32_t in MD5), with uint8_t* input and a size of input. Testbed script uses the library and is designed to allow easy testing of the time taken to perform a X hashes of N length. For memory optimization, large arrays of constants have been moved to Progmem or converted into functions where possible.